new-newpage
This is the last patch before ones that start to allow for different page sizes.
newpage will need to get the right page size from its segment, but the ancient
calling sequence only passes the Segment* when the segment is locked, as a Segment**, and then
if it must be unlocked by newpage, the Segment* it points to is set to nil.

Replace all that by newpage(int, Segment*, uintptr, int locked)
where Segment* is passed in all cases, and
where locked is true if the segment must be unlocked if newpage must sleep,
and newpage will then return nil (otherwise it always returns a valid Page*).

There's also a bug fix to sysseg.c from Plan 9, preventing segments from shrinking underfoot.

Author(s): forsyth

Reference: 016-new-newpage
Date: Wed Mar 13 02:26:27 GMT 2013
Signed-off-by: jmk