segment-param-2
part 1 of making page size possibly variable:
add log2 page size to Page, shrinking modref and color to bytes.
instead of the shifts, it might be better to tidier to index an array,
or use a macro, not least because the type involved is uintptr not int,
but I'll leave it as it is for the moment until some more shuffling
has been done.

Author(s): forsyth

Reference: 019-segment-param-2
Date: Wed Mar 13 16:02:32 GMT 2013
Signed-off-by: jmk