schedinit-reorder
move edfstop and call to free out of Moribund in schedinit,
by doing it all in pexit, after resources are released
(so that's done with edf constraints if active);
and don't updatecpu for Moribund processes *after*
releasing up(!) for re-use. the edfstop/free could go where
it previously was in pexit, after locking procalloc/palloc,
but since all resources, including wait records, have been
released, and pid is 0, and unhashed,
there doesn't seem to be a visible advantage in keeping
edf for the run to schedinit.

Author(s): forsyth

Reference: 021-schedinit-reorder
Date: Wed Mar 13 16:19:46 GMT 2013
Signed-off-by: jmk