etherigbecls8
We had to set cls in this case to boot on virtual box with a bridged
adapter on an imac.
the manual says that values <16 are ignored and taken as 0, but the code
from older kernels was using 8.
We are not sure this is the actual change to make; although it works.

[I made this change a little differently, based on what information is
already at hand in the Pci struct and what was done in other drivers;
made a similar change to 386/ether82563.c;
updated 386/etherigbe.c, 386/ether82598.c to match Plan 9.]

Author(s): nemo, jmk

Reference: 026-etherigbecls8
Date: Thu Mar 14 18:05:06 GMT 2013
Signed-off-by: jmk