tinyfixes0315lsub
This adds some tiny things we fixed while 9k was
being synced.
A comment in mem, do not reset on the spurious intr. we get at clu by now,
print (void) swap info so 9 scripts are happy with /dev/swap, and add
a dependency to mkfile.
Plus, it includes the change to qmalloc indicated by forsyth while
wait wait for an actual fix in qmalloc.

[Pruned some other (I think) unised code from qmalloc.c, which caused
a weird scope problem to show up in devmnt.c concerning the type of the
msize argument to mntversion; punted on deciding the proper type for
now]

Author(s): nemo, jmk

Reference: 028-tinyfixes0315lsub
Date: Wed Mar 20 01:03:40 GMT 2013
Signed-off-by: jmk