waitwhile
Original Readme:
Make waitwhile return instead of waiting forever for the
value to change, and rename it to waitfor so we don't think
it waits while.

We've discussed this a lot. It will likely change once we
find interesting things to do with it, but this has been sitting
for a long time and, I suspect, holding things up. I added code
for the underlying primitives mostly as reference. --jmk

Author(s): nemo

Reference: 043-waitwhile
Date: Thu May 23 14:36:31 GMT 2013
Signed-off-by: jmk